The latest revision improves our treatment of SILK builtins (those predicates having silk:, silkb: and swrlb: predicates) by preventing them from being returned in hilog queries (bug 98) and also by making them non-tabled predicates.  (Justification builtins, which need to be tabled, are an exception to the latter.)  The revision adds additional validation to the OmniRule Transform, eliminating output rules that have negated builtins as heads.  The revision also adds this validation to user rules, raising an exception in this case.  Instances of 'neg <builtin>' in rule bodies are also automatically translated to 'naf <builtin>' internally.  The revision also loads all the external predicate definitions contained in and derived from one file all at once, instead of individually, saving load time.
The implementation of this treatment of SILK builtins deserves some additional commenting here.  All builtin predicates will be defined in the Flora module 'builtin', unlike generic predicates, which are automatically defined in the module 'main'.  This is implemented so as to be entirely transparent to the user, who never uses modules in SILK.  All this is achieved as follows.  All rules and queries containing builtin literals are specially serialized before being sent to Flora.  All such literals are prepended with '%', making the builtin non-tabled in Flora.  All occurrences of builtins in the body have that literal appended with '@builtin'.  In addition, all rules with builtins in the head are automatically loaded into the (Flora) module 'builtin'.  

The new treatment of builtins has broken the persistent query mechanism, which relied on query predicates being tabled.  I attempted a fix which seems like it should work, but does not.  This may be a bug in Flora - stay tuned.

The revision also adds the Flora API and interprolog source code to the SILK execution stack (heretofore, .jar files were used for these components), improving our debugging capabilities.  The revision also fixes bug 127, ensuring that assert and unassert handle quoted content identically.  The revision adds a small fix to CopySILK necessary for windows.  The revision adds getters for prefix and namespace for Vocabulary.  The revision adds some simple profiling capabilities to the engine.  The revision also adds an 'execute' command to the SILK CLI and Engine which feeds an engine command string directly to the implementing engine.



git-svn-id: svn://svn.code.sf.net/p/flora/src/trunk@195 0877669f-6caa-400d-a9c2-f8e882ed57a8